moved all routing to respective children
