+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/channel/routes/channel-routing.module.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { ChannelComponent } from '../../components/channel/channel.component';
+import { ChannelComponent } from '../channel.component';
 console.warn('LOADED CHANNEL !!!!');
 export const channelRoutes: Routes = [
   {
