+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/user-dashboard/customization/routes/edit-routing.module.ts
@@ -1,5 +1,5 @@
 import { Routes } from "@angular/router";
-import { CustomizationComponent } from "../../../components/user-dashboard/customization/customization.component";
+import { CustomizationComponent } from "../customization.component";
 
 export const customizationRoutes: Routes = [
     {
