+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/components/user-dashboard/dashboard/routes/dashboard-routing.module.ts
@@ -1,5 +1,5 @@
 import { Routes } from "@angular/router";
-import { DashboardComponent } from "../../../components/user-dashboard/dashboard/dashboard.component";
+import { DashboardComponent } from "../dashboard.component";
 console.warn("Dashboard was Loaded------------ !!");
 export const dashboardRoutes: Routes = [
   {
