+++ youtube-clone-back-end/frontend/youtube-clone-ui/src/app/module/icons-provider.module.ts
